requirements.txt
python-dotenv
fastapi
uvicorn
qdrant-client
openai
pydantic
requests
pymupdf
langchain-text-splitters